refactor : rename div(div.styled)
